restructure pattern code
